Improve Plus & Input too long dialog
